Documentation & vignette edits in line with new exhaustive and stepwise searches in the paper. MoE_compare now also accepts "MoECompare" objects as (one of the) input(s). Included equalPro in best model printing info. Small fix to drop_levels(). MoE_compare gains argument noise.vol for overriding the noise.meth argument and specifying an improper uniform density directly via the (hyper)volume (which is explicitly necessary & warned about for high-dimensional data). Slight fixes for when tau0 is a vector.
Small MoE_gpairs fixes, especially for models with a noise component: added expert.covar as an option to MoE_gpairs (also to as.Mclust); fixes to subset$data.ind and subset$cov.ind arguments; now works (with a warning) for single panel plots (i.e. univariate data with no covariates); colours are now added to barcode plots for panels involving the MAP classification; MoE_plotGate colours are now consistent with MoE_gpairs.

Properly accounted for predictions of single observations. Properly accounted for predictions for the noise component. This was helped by summarising expert network coefficients for the noise component as though it was an intercept only regression at convergence.