Overhauled the function noise_vol to now also return the d-dimensional location vector giving the centre of mass of the noise component. This works for the range, hypvol, ellipsoidhull, & convexhull.
Thus the fitted values used to summarise the means in the presence of expert covariates and a noise component is modified again. Now the renormalisation approach to correcting for the summarised means in the presence of expert network covariates and a noise component is ONLY used when noise.vol is supplied manually.

The volume for method="ellipsoidhull" is now computed correctly, and also on the log-scale, owing to a bug in the cluster package.

The predict.MoEClust function has been modified to reflect the above. Also, prediction (now including ystar!) for G=0 models has been re-allowed. Prepared new release.